Added steering centre members to VehicleData.
steeringReturn is the main control for the amount of return in
the steering. Setting it to 0 (the default) disables steering return
completely.

steeringReturnSpeedScale is a multiplier on the amount of effect that
the vehicle's velocity has on its steering return.

powerSteering can be set to eliminate the effect of steering return on
user input. Without this flag set, the user will have a harder time
steering, because the wheels try to return as they steer. If this flag
is set, steering return will not be applied while the user is actually
steering, only when they release.