Change ShapeBase to only test if available meshes
The ShapeBase class tests if all of its meshes are hidden and if so,
then it doesn't render during that frame.  However, if there are no
meshes in the ShapeBase and only nodes then rendering is also skipped.
Unfortunately, this also skips the rendering of any mounted images.
This change allows for a mesh-less skeleton to render any mounted
images.