Lit Alpha Skyboxes
This allows skyboxes to be transparent and lit to work with
scatterskies.  The fog band has been improved, being a cylinder with a
smoother top edge.  The skybox also has color and exposure settings.
The advanced clouds are given a simple flag that allows them to be
displayed behind or in front of a scatter sky.

For a normal ( no alpha, not lit) skybox to function as before, the
material needs to be set as emissive.

Example screenshots and a skybox are available here:

http://www.garagegames.com/community/blogs/view/22265