GuiTreeView bug in buildVisibleTree
There was a bug in buildVisibleTree, it should be setting the RebuildVisible flag, not clearing it.

The clip rectangle did not appear to be updating when expanding items or calling scrollVisible, which expands all the items to make a particular item visible. This would cause part of the tree to be cut off.

Setting the RebuildVisible flag made the problem go away.