banding corrections.
script
rendermanager.cs- Like it says at the bottom of the file, PFX_PassthruShader is just that, a bottleneck through which all shaders run. The end result of that is that the lighting buffer, which at a minimum uses 16RBG becomes downsampled and bands. Steps were taken with stock torque to mitigate that, but with linearization incoming, the effect is exacerbated.
script(hlsl/glsl) - the saturate command (clamp in glsl) ensures the results fall into a 0-1 range (you'd get a black circle in the center of the sun otherwise)
source: simplicfication

immediate purpose - slightly noticeable smoothing of light banding with heavily bright lights in extremely dark environments

long term impact - shuts off a future issue with corrected lighting calculations.