Dev forest wind emitter improvement
So the problem is that when your inside the sphere it won't render so it might make someone
think that it's not working right.  So what I did was determine if the camera is inside the sphere.
If the camera is inside the sphere, then I find the distance from the center of the sphere to the camera
Round down and use that as the radius to draw the sphere.
That way if someone zooms in or out, their screen is still showing the sphere.