This patch has two pieces to it due to the way the parts overlap.
First, I've added Far Clipping Client culling to the scene.  In my benchmarks it increases framerate by 5-10 fps in heavily populated zones.
I've added a new global TorqueScript Variable $pref::Camera::FarDistance to allow easy setting and getting of the clipping distance.
I also added it to the Options Dialog for in game use.
The max value the Far Clipping Distance can be is the max Viewable distance defined in the mission.

With this patch, I've also included a fix for the hitching for when players zoom out for the first time or when they enter a heavily populated area.
The idea is to load all LOD's of all Shapebase and Static objects into the video card before the first render.  This removes the hitch players
see when they scroll out after loading into the game.