Fix compiler errors and cleanup of hash interface functions
Since clang and gcc perform proper two-phase name lookup as mandated in the C++
standard, they don't find functions called in a template that are defined after
the template but before the instantiation, except in special circumstances.
Previously this was being worked around by pre-declaring hash functions before
the hashtable definition, but this doesn't work for a unity build without
putting the functions in the same file as the hash table. Fix this the
preferred way by triggering argument dependent lookup.