Methods and improvements for SimObjects
Method for getting the arguments for a script method.
A bunch of methods
for getting static and dynamic field info (if it's static or dynamic,
count, type, name...).
Added console method for checking if the object
is of a class (checks inheritance).
Other little changes like if
something is const, protected vs private...
Changed deleteObject to virtual so it could be used in other classes.