Added Methods to SimSet
Can now find first object of a class.
Get a space delimited list of
object ids in the set (also only of certain class) in script
Get hidden
and selectable (recursive) count.
Separated isMember  out from the
console method so it can be used in engine easily.
Opened reorder method
to script.
Added callbacks for onReOrder and onClear so a SimSet in
script can be notified of changes.