Fix TORQUE_UNUSED for recent versions of MSVC
Since there's now apparently no way to suppress the warning for a
particular variable without adding at least some extra size to the
executable, just turn the warning off in release builds. We leave it on in
debug since it can sometimes help catch bugs, and we don't care about a
little extra code in that configuration.