Improved fading and animation methods
Added getThreadSequence so there wasn't just a setter for it.
Added
getThreadTimeScale so there wasn't just a setter for it.
Added parameter
to startFade for a time to start within the fade so it can start half
faded for example.
Added pauseFade method to pause/resume fading.
Added
isFading method to check if an object is fading out or not.
Fixed some
spacing issues in updateThread method.
Added getThreadAnim to get the
name of the animation for a given thread in script.
Added
getMaxThreadCount script method to get the max number of threads we can
have.
Added getThreadTimeScale method so we can not just set but get the
thread speed/time scale.