Added JSON object
Added a JSONObject  that inherits from ArrayObject so you can create a
json strings.
If you want treat the elements as a JSON array object set isArray to
true and all keys will be ignored.
You can have JSON objects inside JSON
objects by using the name or id of an JSON object as the value.