Split vsync into fullscreen and windowed vsync with "smart" defaults
Previously we were defaulting to vsync=off, but in recent versions
of windows this is often a bad idea because desktop compositing ("aero")
effectively enforces a vsync in windowed modes, regardless of our
presentation interval. So turning on vsync in those cases is beneficial
because we won't waste CPU effort rendering frames that are just going to
be discarded by the OS anyways, and frame rate displays will more
accurately represent what the user is actually seeing on the monitor.

To solve this, the engine now tracks separate fullscreen and windowed
vsync modes. Fullscreen defaults to vsync=off. Windowed modes default to
vsync=on if desktop composition is detected, and off otherwise. These
settings are stored in the vsyncFullscreen and vsyncWindowed config
variables. The settings are:

0 : vsync disabled
1 : vsync enabled
2 : auto vsync (default)

These variables aren't intended to be changed directly, and they don't
appear in the graphic options. Any changes to the vsync setting via the
graphics options will removed the "auto" behavior, and this can only be
restored by resetting or editing the config by hand.

Also, rename the old vsync config variable so it won't get saved.
Its value will be clobbered by the fullscreen and windowed variables
anyway.