PVS Issues found in mEase.h:
The 'EaseF' class implements a copy constructor, but lacks the '=' operator. It is dangerous to use such a class

Undefined behavior. The 't' variable is modified while being used twice between sequence points, this one occurred multiple times

As well as some formatting/cleanup to make it adhere to the code guidelines