Fix fox SFXPlaylist breaking networking
SFXPlaylist tries to write too many bits into the bitstream, which seems
to be responsible for a few issues people have had with non-local
servers. This might not be the best way to go about fixing it, but its
not as bad as what is in the engine right now.