Select camera when in material editor crashfix.
When attempting to select a camera Object(so going into the editor, going into the freefloating camera and then back to the player camera and selecting the freefloating one) while the material editor was open, it would crash because it attempts to access the shapebase object's shapeInstance - but cameras don't have a shape.

This fixes it so it makes sure there's a shapeinstance before trying to get the material data.