Alt+C Camera Toggle Bug Fix
Removed the leftover line of code changing the keybind of the alt+c keyboard command that was supposed to have been moved to a new location, but was not. In the new location, the keyboard command alt+c is bound to a different function that properly toggles the camera as alt+c should.