extra option for terrain block to disable the update of the basetexture
This adds another checkbox in the terrain datablock in the editor, where you can disable the auto generation of the basetexture, which allows you to use a custom basetexture for the whole terrain. Terrain generation software often puts out a precalculated basetexture and with this option you can use it by just renaming it into level.basetex.png or whatever and the engine will use it and not overwrite it.